+++ test/lib/server/index.js
@@ -14,11 +14,11 @@
 const UtilitiesSuite = require('./utilities/index.js');
 const ServicesSuite = require('./services/index.js');
 const MiddlewareSuite = require('./middleware/index.js');
-const MonalisaSynchronizationSuite = require('./MonALISASynchronization/index.js');
+const MonALISASynchronizationSuite = require('./monalisa-synchronization/index.js');
 
 module.exports = () => {
     describe('Utilities', UtilitiesSuite);
     describe('Services', ServicesSuite);
     describe('Middlewares', MiddlewareSuite);
-    describe('External Services Synchronization', MonalisaSynchronizationSuite);
+    describe('External Services Synchronization', MonALISASynchronizationSuite);
 };
